requirements.txt
coreapi==2.0.8
coreschema==0.0.4
Django==1.11.3
django-braces==1.8.1
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-extensions==1.7.9
django-filter==1.0.4
django-rest-swagger==2.1.1
djangorestframework==3.6.3
djangorestframework-jsonapi==2.1.1
google-api-python-client==1.5.1
psycopg2
httplib2==0.9.2
inflection==0.3.1
itypes==1.1.0
mongoengine==0.10.6
oauth2client==2.2.0
oauthlib==1.0.3
openapi-codec==1.1.7
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyJWT==1.4.0
pymongo==3.3.0
python-openid==2.2.5
-e git://github.com/toladata/social-core#egg=social-core
social-auth-app-django==1.2.0
pytz==2015.6
PyYAML==3.11
requests==2.11.1
requests-oauthlib==0.5.0
rsa==3.4.2
selenium==2.48.0
simplejson==3.9.0
six==1.10.0
uritemplate==3.0.0
celery==4.0.2
django-celery-results==1.0.1
